+++ 42cursus/Libft/ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: alperrot <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/12 13:36:29 by alperrot          #+#    #+#             */
-/*   Updated: 2023/11/12 16:04:58 by alperrot         ###   ########.fr       */
+/*   Updated: 2023/11/15 19:47:13 by alperrot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 char	*ft_strjoin(char const *s1, char const*s2)
 {
 	char	*str;
-	int		i;
-	int		j;
+	size_t	i;
+	size_t	j;
 
 	if (!s1 || !s2)
 		return ((void *) 0);
