+++ 42cursus/Libft/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: alperrot <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 16:48:58 by alperrot          #+#    #+#             */
-/*   Updated: 2023/11/12 14:00:29 by alperrot         ###   ########.fr       */
+/*   Updated: 2023/11/15 19:48:05 by alperrot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	int		i;
+	size_t	i;
 
 	if (size == 0)
 		return (ft_strlen(src));
 	i = 0;
-	while (src[i] && (i < (int) size - 1))
+	while (src[i] && (i < size - 1))
 	{
 		dst[i] = src[i];
 		i++;
